+++ cmd/cmd.go
@@ -1,6 +1,11 @@
 package cmd
 
-import "github.com/ghifari160/medhash-tools/color"
+import (
+	"os"
+	"path/filepath"
+
+	"github.com/ghifari160/medhash-tools/color"
+)
 
 const (
 	MsgStatusError   = color.Red + "ERROR" + color.Reset
@@ -8,6 +13,10 @@ const (
 	MsgStatusSkipped = color.Yellow + "SKIPPED" + color.Reset
 	MsgFinalError    = color.Red + "Error!" + color.Reset
 	MsgFinalDone     = color.Green + "Done!" + color.Reset
+
+	PrefixError = color.Red + "Error:" + color.Reset
+
+	ConfDir = "medhash-tools"
 )
 
 // Command is a generic wrapper for all subcommands.
@@ -31,4 +40,18 @@ type CmdConfig struct {
 	SHA256 bool `arg:"--sha256" help:"use SHA256"`
 	SHA1   bool `arg:"--sha1" help:"use SHA1"`
 	MD5    bool `arg:"--md5" help:"use MD5"`
+}
+
+// ConfigDir returns the directory for user configurations.
+func ConfigDir() (dir string, err error) {
+	dir, err = os.UserConfigDir()
+	if err != nil {
+		return
+	}
+
+	dir = filepath.Join(dir, ConfDir)
+
+	err = os.MkdirAll(dir, 0755)
+
+	return
 }